requirements.txt
phidata
python-dotenv
yfinance
duckduckgo_search
groq
streamlit